requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM Providers (install what you need)
anthropic>=0.18.0          # Claude (default)
openai>=1.0.0              # GPT-4, also used for Deepseek
google-generativeai>=0.3.0 # Gemini

# CLI
click>=8.0.0
rich>=13.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Future phases (uncomment when needed)
# chromadb>=0.4.0  # RAG vector store
# pypdf>=3.0.0     # PDF parsing for Bibles
